requirements.txt
requests
googlesearch-python
google